Data.Selection

Description

The recently discovert selection monad provides an elegant way of implementing AI's for sequential games. This library is providing the definition of the selection and quantifier monad as introduced by Martin Escardó and Paulo Oliva in their paper from 2010: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/Escardo-Oliva-MSFP2010.pdf

A detailed introduction to this library as well as some example game implementations can be found in the master thesis of Johannes Hartmann: https://github.com/IncredibleHannes/QMUL-MastersProject

The library consists of the type definitions of the selection and quantifier monad, the definition of the otimes functions, a collection of minimum and maximum functions and a set of functions that compute the optimal play for a given sequential game.

Selection Monad

Definition of the selection monad

Definition is taken from: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/MSFP2010/haskell/modular/monadic/

newtype J r x

Definition of the selection type. The type variable R can be seen as a generalised truth value. For example for r = Bool, a function can be build with the type:

f :: [x] -> K Bool x

that is selecting an element out of the list of x's for wich the given functions

f2 :: (x -> r)

is returning true.

Constructors

J

Fields

selection :: (x -> r) -> x

morphismJK :: J r x -> K r x

Transforing a selection function into a quantifier function

Quantifier Monad

Definition of the quantifier monad

Definition is taken from: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/MSFP2010/haskell/modular/monadic/

newtype K r x

Definition of the quantifier type. The type variable R can be seen as a generalised truth value. For example for r = Bool, a function can be build with the type:

f :: [x] -> K Bool x

is returning true if an element exists for wich the function:

f2 :: (x -> r)

is returning true.

Constructors

K

Fields

quantifier :: (x -> r) -> r

Bigotimes

Definition of the selection monad

Definition is taken from: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/MSFP2010/haskell/modular/monadic/

varotimes :: Monad m => m x -> (x -> m [x]) -> m [x]

otimes :: Monad m => m x -> (x -> m y) -> m (x, y)

bigotimes :: Monad m => [[x] -> m x] -> m [x]

Min|max functions

Definition of different minimum and maximum functions

Generic

epsilonMin :: Ord b => [a] -> J b a

epsilonMax :: Ord b => [a] -> J b a

Win|Draw|Lose

epsilonMinThree :: [a] -> J Three a

epsilonMaxThree :: [a] -> J Three a

Booleans

epsilonMinBool :: [a] -> J Bool a

epsilonMaxBool :: [a] -> J Bool a

Length tracking tuples

epsilonMinTuple :: [a] -> J (Int, Int) a

epsilonMaxTuple :: [a] -> J (Int, Int) a

Paralell versions

Generic

epsilonMaxParalell :: (NFData a, NFData b, Ord a, Ord b) => [b] -> J a b

epsilonMinParalell :: (NFData a, NFData b, Ord a, Ord b) => [b] -> J a b

Length tracking tuples

epsilonMaxTupleParalell :: NFData a => [a] -> J (Int, Int) a

epsilonMinTupleParalell :: NFData a => [a] -> J (Int, Int) a

Optimal game strategies

Definition of the functions that calculate the optimal play

optimalPlay :: ([b] -> a) -> [[b] -> J a b] -> [b]

optimalOutcome :: ([b] -> a) -> [[b] -> J a b] -> a

optimalStrategy :: ([b] -> a) -> [[b] -> J a b] -> [b] -> b