Data.Selection.J

Description

Definition of the selection monad

Definition is taken from: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/MSFP2010/haskell/modular/monadic/

Documentation

newtype J r x

Definition of the selection type. The type variable R can be seen as a generalised truth value. For example for r = Bool, a function can be build with the type:

f :: [x] -> K Bool x

that is selecting an element out of the list of x's for wich the given functions

f2 :: (x -> r)

is returning true.

Constructors

J

Fields

selection :: (x -> r) -> x

unitJ :: x -> J r x

implementing the return function of the monad definition

functorJ :: (x -> y) -> J r x -> J r y

muJ :: J r (J r x) -> J r x

morphismJK :: J r x -> K r x

Transforing a selection function into a quantifier function