Data.Selection.K

Description

Definition of the quantifier monad

Definition is taken from: https://www.cs.bham.ac.uk/~mhe/papers/msfp2010/MSFP2010/haskell/modular/monadic/

Documentation

newtype K r x

Definition of the quantifier type. The type variable R can be seen as a generalised truth value. For example for r = Bool, a function can be build with the type:

f :: [x] -> K Bool x

is returning true if an element exists for wich the function:

f2 :: (x -> r)

is returning true.

Constructors

K

Fields

quantifier :: (x -> r) -> r

unitK :: x -> K r x

functorK :: (x -> y) -> K r x -> K r y

muK :: K r (K r x) -> K r x